Add the sync step before uploading to S3
Before uploading, we now sync our local storage of
historical benchmark data with the data that is stored
in S3. The data in S3 is updated after every benchmark
test is completed on teamcity.

In addition, this commit now checks to make sure
environment variables are set throughout the
codebase for quicker debugging.