Remove geometric means page
The geometric means page was originally created to
get a vague idea of whether or not the changes
we were making to cockroach were improving or
hurting performance. However, since the benchmark
tests often change comparing the "geometric mean of all
test results" is completely useless. This page
could possibly be one day revived if we use a set
of tests that never change.